Address feedback - move big function in extra helper class
