+++ src/modules/launcher/Plugins/Microsoft.PowerToys.Run.Plugin.WindowsSettings/Main.cs
@@ -4,61 +4,55 @@
 
 using System;
 using System.Collections.Generic;
-using System.IO;
 using System.Linq;
 using System.Reflection;
-using System.Text.Json;
-using System.Text.Json.Serialization;
 using ManagedCommon;
 using Microsoft.PowerToys.Run.Plugin.WindowsRegistry.Helper;
 using Microsoft.PowerToys.Run.Plugin.WindowsSettings.Classes;
 using Microsoft.PowerToys.Run.Plugin.WindowsSettings.Helper;
 using Microsoft.PowerToys.Run.Plugin.WindowsSettings.Properties;
 using Wox.Plugin;
-using Wox.Plugin.Logger;
 
 namespace Microsoft.PowerToys.Run.Plugin.WindowsSettings
 {
     /// <summary>
-    /// Main class of this plugin that implement all used interfaces
+    /// Main class of this plugin that implement all used interfaces.
     /// </summary>
     public class Main : IPlugin, IContextMenu, IPluginI18n, IDisposable
     {
         /// <summary>
-        /// The path to the symbol for a light theme
+        /// The path to the symbol for a light theme.
         /// </summary>
         private const string _lightSymbol = "Images/WindowsSettings.light.png";
 
         /// <summary>
-        /// The path to the symbol for a dark theme
+        /// The path to the symbol for a dark theme.
         /// </summary>
         private const string _darkSymbol = "Images/WindowsSettings.dark.png";
 
         /// <summary>
-        /// The name of the file that contains all settings for the query
-        /// </summary>
-        private const string _settingsFile = "WindowsSettings.json";
-
-        /// <summary>
-        /// The name of this assembly
+        /// The name of this assembly.
         /// </summary>
         private readonly string _assemblyName;
 
         /// <summary>
-        /// The initial context for this plugin (contains API and meta-data)
+        /// The initial context for this plugin (contains API and meta-data).
         /// </summary>
         private PluginInitContext? _context;
 
         /// <summary>
-        /// The path to the icon for each result
+        /// The path to the icon for each result.
         /// </summary>
         private string _defaultIconPath;
 
         /// <summary>
-        /// Indicate that the plugin is disposed
+        /// Indicate that the plugin is disposed.
         /// </summary>
         private bool _disposed;
 
+        /// <summary>
+        /// List that contain all settings.
+        /// </summary>
         private IEnumerable<WindowsSetting>? _settingsList;
 
         /// <summary>
@@ -81,50 +75,27 @@ namespace Microsoft.PowerToys.Run.Plugin.WindowsSettings
         public string Description => Resources.PluginDescription;
 
         /// <summary>
-        /// Initialize the plugin with the given <see cref="PluginInitContext"/>
+        /// Initialize the plugin with the given <see cref="PluginInitContext"/>.
         /// </summary>
-        /// <param name="context">The <see cref="PluginInitContext"/> for this plugin</param>
+        /// <param name="context">The <see cref="PluginInitContext"/> for this plugin.</param>
         public void Init(PluginInitContext context)
         {
             _context = context ?? throw new ArgumentNullException(nameof(context));
             _context.API.ThemeChanged += OnThemeChanged;
             UpdateIconPath(_context.API.GetCurrentTheme());
 
-            var assembly = Assembly.GetExecutingAssembly();
-            var type = assembly.GetTypes().FirstOrDefault(x => x.Name == nameof(Main));
+            var settingsList = ResourceHelper.ReadAllPossibleSettings();
 
-            var resourceName = $"{type?.Namespace}.{_settingsFile}";
+            TranslationHelper.TranslateAllSettings(settingsList);
 
-            try
-            {
-                using var stream = assembly.GetManifestResourceStream(resourceName);
-                if (stream is null)
-                {
-                    throw new Exception("stream is null");
-                }
-
-                var options = new JsonSerializerOptions();
-                options.Converters.Add(new JsonStringEnumConverter());
-
-                using var reader = new StreamReader(stream);
-                var text = reader.ReadToEnd();
-
-                _settingsList = JsonSerializer.Deserialize<IEnumerable<WindowsSetting>>(text, options);
-            }
-            catch (Exception exception)
-            {
-                Log.Exception("Error loading settings JSON file", exception, typeof(Main));
-            }
-
-            TranslationHelper.TranslateAllSettings(_settingsList);
-            _settingsList = WindowsVersionHelper.FilterByVersion(_settingsList);
+            _settingsList = WindowsVersionHelper.FilterByVersion(settingsList);
         }
 
         /// <summary>
-        /// Return a filtered list, based on the given query
+        /// Return a filtered list, based on the given query.
         /// </summary>
-        /// <param name="query">The query to filter the list</param>
-        /// <returns>A filtered list, can be empty when nothing was found</returns>
+        /// <param name="query">The query to filter the list.</param>
+        /// <returns>A filtered list, can be empty when nothing was found.</returns>
         public List<Result> Query(Query query)
         {
             if (_settingsList is null)
@@ -169,10 +140,10 @@ namespace Microsoft.PowerToys.Run.Plugin.WindowsSettings
         }
 
         /// <summary>
-        /// Return a list context menu entries for a given <see cref="Result"/> (shown at the right side of the result)
+        /// Return a list context menu entries for a given <see cref="Result"/> (shown at the right side of the result).
         /// </summary>
-        /// <param name="selectedResult">The <see cref="Result"/> for the list with context menu entries</param>
-        /// <returns>A list context menu entries</returns>
+        /// <param name="selectedResult">The <see cref="Result"/> for the list with context menu entries.</param>
+        /// <returns>A list context menu entries.</returns>
         public List<ContextMenuResult> LoadContextMenus(Result selectedResult)
         {
             return ContextMenuHelper.GetContextMenu(selectedResult, _assemblyName);
@@ -186,9 +157,9 @@ namespace Microsoft.PowerToys.Run.Plugin.WindowsSettings
         }
 
         /// <summary>
-        /// Wrapper method for <see cref="Dispose"/> that dispose additional objects and events form the plugin itself
+        /// Wrapper method for <see cref="Dispose"/> that dispose additional objects and events form the plugin itself.
         /// </summary>
-        /// <param name="disposing">Indicate that the plugin is disposed</param>
+        /// <param name="disposing">Indicate that the plugin is disposed.</param>
         protected virtual void Dispose(bool disposing)
         {
             if (_disposed || !disposing)
@@ -221,19 +192,19 @@ namespace Microsoft.PowerToys.Run.Plugin.WindowsSettings
         }
 
         /// <summary>
-        /// Change all theme-based elements (typical called when the plugin theme has changed)
+        /// Change all theme-based elements (typical called when the plugin theme has changed).
         /// </summary>
-        /// <param name="oldtheme">The old <see cref="Theme"/></param>
-        /// <param name="newTheme">The new <see cref="Theme"/></param>
+        /// <param name="oldtheme">The old <see cref="Theme"/>.</param>
+        /// <param name="newTheme">The new <see cref="Theme"/>.</param>
         private void OnThemeChanged(Theme oldtheme, Theme newTheme)
         {
             UpdateIconPath(newTheme);
         }
 
         /// <summary>
-        /// Update all icons (typical called when the plugin theme has changed)
+        /// Update all icons (typical called when the plugin theme has changed).
         /// </summary>
-        /// <param name="theme">The new <see cref="Theme"/> for the icons</param>
+        /// <param name="theme">The new <see cref="Theme"/> for the icons.</param>
         private void UpdateIconPath(Theme theme)
         {
             _defaultIconPath = theme == Theme.Light || theme == Theme.HighContrastWhite
